+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Fixed
+* Corrected bug in statistics paging strategy
+* Exposed `getObjectIdRange` as public function
+
 ## [1.5.4] - 2016-03-30
 ### Fixed
 * Don't throw exception when passed in layer is of type numeric